Rename profileList to profilelist to match naming conventions
